Add quote block, call to action, footer
I spent a bit of time getting all three of these sections done in
addition to going back and cleaning up the rest of the page
before sending this final commit and pushing back to the remote.

Though it's rather hacky in some places, it's a good starting
place and I'm proud of the work that went into it.